requirements.txt
streamlit
python-docx
python-pptx
PyMuPDF
google-generativeai
dotenv
streamlit-agraph
graphviz
pyvis
